+++ js/main.js
@@ -1,68 +1,4 @@
-const COMMENTS_SENTENCES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-const NAMES = [
-  'Иван',
-  'Владимр',
-  'Александр',
-  'Вадим',
-  'Илья',
-  'Максим',
-  'Яна',
-  'Анна',
-  'Алина',
-  'Злата',
-];
-
-const getRandomInteger = (min, max) => {
-  const lower = Math.ceil(Math.min(min, max));
-  const upper = Math.floor(Math.max(min, max));
-  return Math.floor(Math.random() * (upper - lower + 1) + lower);
-};
-
-const createRandomIdFromRangeGenerator = (min, max) => {
-  const previousValues = [];
-  return function () {
-    let currentValue = getRandomInteger(min, max);
-    if (previousValues.length >= (max - min + 1)) {
-      return null;
-    }
-    while (previousValues.includes(currentValue)) {
-      currentValue = getRandomInteger(min, max);
-    }
-    previousValues.push(currentValue);
-    return currentValue;
-  };
-};
-
-const getRandomElementFromArray = (arr) => arr[getRandomInteger(0, arr.length - 1)];
-
-const PHOTO_COUNT = 25;
-const generatePhotoId = createRandomIdFromRangeGenerator(1, 25);
-const generateCommentId = createRandomIdFromRangeGenerator(1, 750);
-
-const createComment = () => ({
-  id: generateCommentId(),
-  avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
-  message: getRandomElementFromArray(COMMENTS_SENTENCES),
-  name: getRandomElementFromArray(NAMES),
-});
-
-const createPhoto = () => {
-  const photoID = generatePhotoId();
-  return {
-    id: photoID,
-    url: `photos/${photoID}.jpg`,
-    description: `Описание фото номер ${photoID}`,
-    likes: getRandomInteger(15, 200),
-    comments: Array.from({length: getRandomInteger(0, 30)}, createComment),
-  };
-};
+import {createPhotos} from '.data.js';
 
 // eslint-disable-next-line no-unused-vars
-const photos = Array.from({length: PHOTO_COUNT}, createPhoto);
+const photos = createPhotos();
